Fixes Site ID mismatches in a users OneDrive URL
Often occurs on a few unlucky users after a tenancy migration. The standard format for OneDrive is: https://(SITE ID)-my.sharepoint.com/personal/(UPN) however the Site ID can be defined incorrectly as the old tenants site ID. This script is compares the Site ID found of the specified user and if it doesn't match the one found for the tenant, it replaces it. Currently untested, and potentially can be automated further.